# Remove scheduler in-flight resources dict in favor of transactions