--- a/internal/scheduler/legacyscheduler.go
+++ b/internal/scheduler/legacyscheduler.go
@@ -160,11 +160,6 @@ func (it *QueueCandidateJobsIterator) Lease(jobSchedulingReport *JobSchedulingRe
 	it.totalQueueResources = jobSchedulingReport.TotalQueueResources
 	it.roundQueueResources = jobSchedulingReport.RoundQueueResources
 	it.totalQueueResourcesByPriority = jobSchedulingReport.TotalQueueResourcesByPriority
-	if it.NodeDb != nil {
-		for _, report := range jobSchedulingReport.PodSchedulingReports {
-			it.NodeDb.BindNodeToPod(jobSchedulingReport.JobId, report.Req, report.Node)
-		}
-	}
 }
 
 func (it *QueueCandidateJobsIterator) Next() (*JobSchedulingReport, error) {
@@ -394,7 +389,7 @@ func (sched *LegacyScheduler) jobIsLargeEnough(jobTotalResourceRequests common.C
 	return true, ""
 }
 
-func (sched *LegacyScheduler) selectNodeForPod(ctx context.Context, jobId uuid.UUID, job *api.Job) (*PodSchedulingReport, error) {
+func (sched *LegacyScheduler) selectNodeForPod(ctx context.Context, jobId uuid.UUID, job *api.Job, bind bool) (*PodSchedulingReport, error) {
 	podSpec := podSpecFromJob(job)
 	if podSpec == nil {
 		return nil, errors.New("failed to get pod spec")
@@ -403,7 +398,13 @@ func (sched *LegacyScheduler) selectNodeForPod(ctx context.Context, jobId uuid.U
 	// Try to find a node for this pod.
 	// Store the report returned by the NodeDb.
 	req := schedulerobjects.PodRequirementsFromPodSpec(podSpec, sched.SchedulingConfig.Preemption.PriorityClasses)
-	report, err := sched.NodeDb.SelectNodeForPod(jobId, req)
+	var report *PodSchedulingReport
+	var err error
+	if bind {
+		report, err = sched.NodeDb.SelectAndBindNodeToPod(jobId, req)
+	} else {
+		report, err = sched.NodeDb.SelectNodeForPod(jobId, req)
+	}
 	if err != nil {
 		return nil, err
 	}
@@ -642,7 +643,7 @@ func (sched *LegacyScheduler) Schedule(
 			// if other jobs were scheduled onto that node in the interim.
 			//
 			// TODO: Only repeat this process if the node in the report no longer works.
-			podReport, err := sched.selectNodeForPod(ctx, report.JobId, report.Job)
+			podReport, err := sched.selectNodeForPod(ctx, report.JobId, report.Job, true)
 			if err != nil {
 				return nil, err
 			}