--- a/internal/scheduler/nodedb.go
+++ b/internal/scheduler/nodedb.go
@@ -7,6 +7,7 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/golang/protobuf/proto"
 	"github.com/google/uuid"
 	"github.com/hashicorp/go-memdb"
 	"github.com/pkg/errors"
@@ -48,20 +49,11 @@ type NodeDb struct {
 	indexedNodeLabels map[string]interface{}
 	// Total amount of resources, e.g., "cpu", "memory", "gpu", managed by the scheduler.
 	// Computed approximately by periodically scanning all nodes in the db.
-	totalResources map[string]resource.Quantity
+	totalResources schedulerobjects.ResourceList
 	// Set of node types. Populated automatically as nodes are inserted.
 	// Node types are not cleaned up if all nodes of that type are removed from the NodeDb.
 	NodeTypes map[string]*schedulerobjects.NodeType
-	// Resources allocated by the scheduler to in-flight jobs,
-	// i.e., jobs for which resource usage is not yet reported by the executor.
-	inFlightResourcesByNode map[string]schedulerobjects.AllocatedByPriorityAndResourceType
-	// Map from job id to the set of nodes on which that job has been assigned resources.
-	// Used to clear AssignedByNode once jobs start running.
-	NodesByJob map[uuid.UUID]map[string]interface{}
-	// Map from node id to the set of jobs that have resourced assigned to them on that node.
-	// Used to clear AssignedByNode once jobs start running.
-	JobsByNode map[string]map[uuid.UUID]interface{}
-	// Mutex to control access to AssignedByNode, NodesByJob, and JobsByNode.
+	// Mutex to control access to totalResources and NodeTypes.
 	mu sync.Mutex
 }
 
@@ -80,16 +72,13 @@ func NewNodeDb(priorities []int32, indexedResources, indexedTaints, indexedNodeL
 		return rv
 	}
 	return &NodeDb{
-		priorities:              priorities,
-		indexedResources:        mapFromSlice(indexedResources),
-		indexedTaints:           mapFromSlice(indexedTaints),
-		indexedNodeLabels:       mapFromSlice(indexedNodeLabels),
-		NodeTypes:               make(map[string]*schedulerobjects.NodeType),
-		totalResources:          make(map[string]resource.Quantity),
-		Db:                      db,
-		NodesByJob:              make(map[uuid.UUID]map[string]interface{}),
-		JobsByNode:              make(map[string]map[uuid.UUID]interface{}),
-		inFlightResourcesByNode: make(map[string]schedulerobjects.AllocatedByPriorityAndResourceType),
+		priorities:        priorities,
+		indexedResources:  mapFromSlice(indexedResources),
+		indexedTaints:     mapFromSlice(indexedTaints),
+		indexedNodeLabels: mapFromSlice(indexedNodeLabels),
+		NodeTypes:         make(map[string]*schedulerobjects.NodeType),
+		totalResources:    schedulerobjects.ResourceList{Resources: make(map[string]resource.Quantity)},
+		Db:                db,
 	}, nil
 }
 
@@ -108,32 +97,43 @@ func (nodeDb *NodeDb) String() string {
 			fmt.Fprintf(w, "  %s\n", nodeType.Id)
 		}
 	}
-	nodeDb.mu.Lock()
-	fmt.Fprintf(w, "Number of in-flight jobs:\t%d\n", len(nodeDb.NodesByJob))
-	nodeDb.mu.Unlock()
 	w.Flush()
 	return sb.String()
 }
 
 func (nodeDb *NodeDb) SelectAndBindNodeToPod(jobId uuid.UUID, req *schedulerobjects.PodRequirements) (*PodSchedulingReport, error) {
-	report, err := nodeDb.SelectNodeForPod(jobId, req)
+	txn := nodeDb.Db.Txn(true)
+	defer txn.Abort()
+	report, err := nodeDb.SelectAndBindNodeToPodWithTxn(txn, jobId, req)
 	if err != nil {
 		return nil, err
 	}
+	txn.Commit()
+	return report, nil
+}
 
+func (nodeDb *NodeDb) SelectAndBindNodeToPodWithTxn(txn *memdb.Txn, jobId uuid.UUID, req *schedulerobjects.PodRequirements) (*PodSchedulingReport, error) {
+	report, err := nodeDb.SelectNodeForPodWithTxn(txn, jobId, req)
+	if err != nil {
+		return nil, err
+	}
 	if report.Node != nil {
-		err = nodeDb.BindNodeToPod(jobId, req, report.Node)
+		err = nodeDb.BindNodeToPod(txn, jobId, req, report.Node)
 		if err != nil {
 			return nil, err
 		}
 	}
-
 	return report, nil
+}
+
+func (nodeDb *NodeDb) SelectNodeForPod(jobId uuid.UUID, req *schedulerobjects.PodRequirements) (*PodSchedulingReport, error) {
+	return nodeDb.SelectNodeForPodWithTxn(nodeDb.Db.Txn(false), jobId, req)
 }
 
 // SelectAndBindNodeToPod selects a node on which the pod can be scheduled,
 // and updates the internal state of the db to indicate that this pod is bound to that node.
-func (nodeDb *NodeDb) SelectNodeForPod(jobId uuid.UUID, req *schedulerobjects.PodRequirements) (*PodSchedulingReport, error) {
+func (nodeDb *NodeDb) SelectNodeForPodWithTxn(txn *memdb.Txn, jobId uuid.UUID, req *schedulerobjects.PodRequirements) (*PodSchedulingReport, error) {
+
 	// Collect all node types that could potentially schedule the pod.
 	nodeTypes, numExcludedNodeTypesByReason, err := nodeDb.NodeTypesMatchingPod(req)
 	if err != nil {
@@ -160,7 +160,6 @@ func (nodeDb *NodeDb) SelectNodeForPod(jobId uuid.UUID, req *schedulerobjects.Po
 	}
 
 	// Iterate over candidate nodes.
-	txn := nodeDb.Db.Txn(false)
 	it, err := NewNodeTypesResourceIterator(
 		txn,
 		dominantResourceType,
@@ -178,9 +177,7 @@ func (nodeDb *NodeDb) SelectNodeForPod(jobId uuid.UUID, req *schedulerobjects.Po
 			break
 		}
 		// TODO: Use the score when selecting a node.
-		nodeDb.mu.Lock()
-		matches, score, reason, err := node.PodRequirementsMet(req, nodeDb.inFlightResourcesByNode[node.Id])
-		nodeDb.mu.Unlock()
+		matches, score, reason, err := node.PodRequirementsMet(req)
 		if err != nil {
 			return nil, err
 		}
@@ -196,36 +193,30 @@ func (nodeDb *NodeDb) SelectNodeForPod(jobId uuid.UUID, req *schedulerobjects.Po
 	return report, nil
 }
 
-func (nodeDb *NodeDb) BindNodeToPod(jobId uuid.UUID, req *schedulerobjects.PodRequirements, node *schedulerobjects.Node) error {
-	nodeDb.mu.Lock()
-	defer nodeDb.mu.Unlock()
+func (nodeDb *NodeDb) BindNodeToPod(txn *memdb.Txn, jobId uuid.UUID, req *schedulerobjects.PodRequirements, node *schedulerobjects.Node) error {
 
-	// Record which jobs have resources assigned to them on each node.
-	if m, ok := nodeDb.JobsByNode[node.Id]; ok {
-		m[jobId] = true
-	} else {
-		nodeDb.JobsByNode[node.Id] = map[uuid.UUID]interface{}{jobId: true}
+	// DeepCopy the node.
+	// TODO: Use more efficient deepcopy.
+	buffer, err := node.Marshal()
+	if err != nil {
+		return err
+	}
+	node = &schedulerobjects.Node{}
+	err = proto.Unmarshal(buffer, node)
+	if err != nil {
+		return err
 	}
 
-	// Record which nodes each job has resources assigned to on.
-	if m, ok := nodeDb.NodesByJob[jobId]; ok {
-		m[node.Id] = true
-	} else {
-		nodeDb.NodesByJob[jobId] = map[string]interface{}{node.Id: true}
-	}
+	// Mark resources as allocated.
+	schedulerobjects.AllocatableByPriorityAndResourceType(node.AllocatableByPriorityAndResource).MarkAllocated(
+		req.Priority,
+		schedulerobjects.ResourceListFromV1ResourceList(req.ResourceRequirements.Requests),
+	)
 
-	// Mark these resources as used.
-	// TODO: Avoid unnecessary copy of req.ResourceRequirements.Requests.
-	rs := schedulerobjects.ResourceList{Resources: make(map[string]resource.Quantity)}
-	for resource, quantity := range req.ResourceRequirements.Requests {
-		rs.Resources[string(resource)] = quantity
-	}
-	if assigned, ok := nodeDb.inFlightResourcesByNode[node.Id]; ok {
-		assigned.MarkAllocated(req.Priority, rs)
-	} else {
-		assigned = schedulerobjects.NewAllocatedByPriorityAndResourceType(nodeDb.priorities)
-		assigned.MarkAllocated(req.Priority, rs)
-		nodeDb.inFlightResourcesByNode[node.Id] = assigned
+	// Insert the node.
+	err = txn.Insert("nodes", node)
+	if err != nil {
+		return errors.WithStack(err)
 	}
 
 	return nil
@@ -266,7 +257,7 @@ func (nodeDb *NodeDb) dominantResource(req *schedulerobjects.PodRequirements) st
 		if _, ok := nodeDb.indexedResources[string(t)]; !ok {
 			continue
 		}
-		available, ok := nodeDb.totalResources[string(t)]
+		available, ok := nodeDb.totalResources.Resources[string(t)]
 		if !ok {
 			return string(t)
 		}
@@ -278,23 +269,6 @@ func (nodeDb *NodeDb) dominantResource(req *schedulerobjects.PodRequirements) st
 		}
 	}
 	return dominantResourceType
-}
-
-// MarkJobRunning notifies the node db that this job is now running.
-// When the nodes were bound to the job, resources on those nodes were marked as assigned in the node db.
-// When the job is running, those resources are accounted for by the executor,
-// and should no longer be marked as assigned in the node db.
-//
-// TODO: This only clears AssignedByNode once there are no in-flight jobs for that node.
-// We could improve it to clear AssignedByNode on a per-job basis.
-func (nodeDb *NodeDb) MarkJobRunning(jobId uuid.UUID) {
-	for nodeId := range nodeDb.NodesByJob[jobId] {
-		delete(nodeDb.JobsByNode[nodeId], jobId)
-		if len(nodeDb.JobsByNode[nodeId]) == 0 {
-			delete(nodeDb.inFlightResourcesByNode, nodeId)
-		}
-	}
-	delete(nodeDb.NodesByJob, jobId)
 }
 
 // Upsert nodes.
@@ -315,18 +289,23 @@ func (nodeDb *NodeDb) Upsert(nodes []*schedulerobjects.Node) error {
 		node.NodeType = nodeType
 
 		// Record all unique node types.
+		nodeDb.mu.Lock()
 		nodeDb.NodeTypes[nodeType.Id] = nodeType
+		nodeDb.mu.Unlock()
 
 		// If this is a new node, increase the overall resource count.
-		if _, ok := nodeDb.inFlightResourcesByNode[node.Id]; !ok {
-			for t, q := range node.TotalResources.Resources {
-				available := nodeDb.totalResources[t]
-				available.Add(q)
-				nodeDb.totalResources[t] = available
-			}
+		it, err := nodeDb.Db.Txn(false).Get("nodes", "id", node.Id)
+		if err != nil {
+			return errors.WithStack(err)
+		}
+		if it.Next() == nil {
+			nodeDb.mu.Lock()
+			nodeDb.totalResources.Add(node.TotalResources)
+			nodeDb.mu.Unlock()
 		}
 
-		err := txn.Insert("nodes", node)
+		// Add the node to the db.
+		err = txn.Insert("nodes", node)
 		if err != nil {
 			return errors.WithStack(err)
 		}