--- a/internal/scheduler/nodedb_test.go
+++ b/internal/scheduler/nodedb_test.go
@@ -562,58 +562,30 @@ func BenchmarkUpsert1(b *testing.B)      { benchmarkUpsert(1, b) }
 func BenchmarkUpsert1000(b *testing.B)   { benchmarkUpsert(1000, b) }
 func BenchmarkUpsert100000(b *testing.B) { benchmarkUpsert(100000, b) }
 
-func benchmarkSelectAndBindNodeToPod(
-	numCpuNodes, numTaintedCpuNodes, numGpuNodes,
-	numSmallCpuJobsToSchedule, numLargeCpuJobsToSchedule, numGpuJobsToSchedule int,
-	b *testing.B,
-) {
+func benchmarkSelectAndBindNodeToPod(nodes []*schedulerobjects.Node, reqs []*schedulerobjects.PodRequirements, b *testing.B) {
 	db, err := NewNodeDb(testPriorities, testResources)
 	if !assert.NoError(b, err) {
 		return
 	}
-	nodes := testNodes3(numCpuNodes, numTaintedCpuNodes, numGpuNodes, testPriorities)
+
 	err = db.Upsert(nodes)
 	if !assert.NoError(b, err) {
 		return
 	}
 
-	smallCpuJob := testSmallCpuJob(0)
-	largeCpuJob := testLargeCpuJob(0)
-	gpuJob := testGpuJob(0)
+	// Pre-allocate job ids.
+	// Re-using these between benchmark iterations is safe
+	// since we clear the jobs between iterations.
+	jobIds := make([]uuid.UUID, len(reqs))
+	for i := range jobIds {
+		jobIds[i] = uuid.New()
+	}
 
 	b.ResetTimer()
 	for n := 0; n < b.N; n++ {
-		jobIds := make([]uuid.UUID, 0)
-		for i := 0; i < numSmallCpuJobsToSchedule; i++ {
-			jobId := uuid.New()
-			jobIds = append(jobIds, jobId)
-			report, err := db.SelectAndBindNodeToPod(jobId, smallCpuJob)
+		for i, req := range reqs {
+			_, err := db.SelectAndBindNodeToPod(jobIds[i], req)
 			if !assert.NoError(b, err) {
-				return
-			}
-			if !assert.NotNil(b, report.Node) {
-				return
-			}
-		}
-		for i := 0; i < numLargeCpuJobsToSchedule; i++ {
-			jobId := uuid.New()
-			jobIds = append(jobIds, jobId)
-			report, err := db.SelectAndBindNodeToPod(jobId, largeCpuJob)
-			if !assert.NoError(b, err) {
-				return
-			}
-			if !assert.NotNil(b, report.Node) {
-				return
-			}
-		}
-		for i := 0; i < numGpuJobsToSchedule; i++ {
-			jobId := uuid.New()
-			jobIds = append(jobIds, jobId)
-			report, err := db.SelectAndBindNodeToPod(jobId, gpuJob)
-			if !assert.NoError(b, err) {
-				return
-			}
-			if !assert.NotNil(b, report.Node) {
 				return
 			}
 		}
@@ -623,6 +595,43 @@ func benchmarkSelectAndBindNodeToPod(
 			db.MarkJobRunning(jobId)
 		}
 	}
+}
+
+func BenchmarkSelectAndBindNodeToPodOneCpuNode(b *testing.B) {
+	benchmarkSelectAndBindNodeToPod(
+		testNCpuNode(1, testPriorities),
+		testNSmallCpuJob(0, 32),
+		b,
+	)
+}
+
+func BenchmarkSelectAndBindNodeToPod100CpuNodes(b *testing.B) {
+	benchmarkSelectAndBindNodeToPod(
+		testNCpuNode(100, testPriorities),
+		testNSmallCpuJob(0, 320),
+		b,
+	)
+}
+
+func BenchmarkSelectAndBindNodeToPod10000CpuNodes(b *testing.B) {
+	benchmarkSelectAndBindNodeToPod(
+		testNCpuNode(10000, testPriorities),
+		testNSmallCpuJob(0, 32000),
+		b,
+	)
+}
+
+func BenchmarkSelectAndBindNodeToPodResourceConstrained(b *testing.B) {
+	nodes := append(append(
+		testNCpuNode(500, testPriorities),
+		testNGpuNode(1, testPriorities)...),
+		testNCpuNode(499, testPriorities)...,
+	)
+	benchmarkSelectAndBindNodeToPod(
+		nodes,
+		testNGPUJob(0, 1),
+		b,
+	)
 }
 
 func testNSmallCpuJob(priority int32, n int) []*schedulerobjects.PodRequirements {
@@ -725,15 +734,3 @@ func testGpuJob(priority int32) *schedulerobjects.PodRequirements {
 // 		},
 // 	}
 // }
-
-func BenchmarkSelectAndBindNodeToPod100(b *testing.B) {
-	benchmarkSelectAndBindNodeToPod(70, 20, 10, 7, 2, 1, b)
-}
-
-func BenchmarkSelectAndBindNodeToPod1000(b *testing.B) {
-	benchmarkSelectAndBindNodeToPod(700, 200, 100, 70, 20, 10, b)
-}
-
-func BenchmarkSelectAndBindNodeToPod10000(b *testing.B) {
-	benchmarkSelectAndBindNodeToPod(7000, 2000, 1000, 700, 200, 100, b)
-}