Simplifies the logic and improves performance.
```bash
# Original
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSelectAndBindNodeToPod10000CpuNodes1CpuUnused$ github.com/G-Research/armada/internal/scheduler
goos: linux
goarch: amd64
pkg: github.com/G-Research/armada/internal/scheduler
cpu: Intel(R) Xeon(R) CPU E5-2697 v3 @ 2.60GHz
BenchmarkSelectAndBindNodeToPod10000CpuNodes1CpuUnused-56    	       1	108566865682 ns/op	22837858544 B/op	400565751 allocs/op
PASS
ok  	github.com/G-Research/armada/internal/scheduler	109.307s

# With this PR
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSelectAndBindNodeToPod10000CpuNodes1CpuUnused$ github.com/G-Research/armada/internal/scheduler
goos: linux
goarch: amd64
pkg: github.com/G-Research/armada/internal/scheduler
cpu: Intel(R) Xeon(R) CPU E5-2697 v3 @ 2.60GHz
BenchmarkSelectAndBindNodeToPod10000CpuNodes1CpuUnused-56    	       1	1097861260 ns/op	265261200 B/op	 5001578 allocs/op
PASS
ok  	github.com/G-Research/armada/internal/scheduler	1.845s
```

That's about a 2 orders of magnitude improvement (from about 100s to about 1s).